requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.8.1
django-ckeditor==4.5.0
django-clever-selects==0.6.9
django-environ==0.3.0
django-extensions==1.5.5
django-filter==1.0.4
django-hosts==2.0
django-localflavor==1.1
django-piwik==0.1
django-rest-swagger==2.1.2
django-smart-selects==1.1.1
django-widget-tweaks==1.3
django-wkhtmltopdf==2.0.2
djangorestframework==3.6.4
djangorestframework-csv==2.0.0
djangorestframework-xml==1.3.0
flake8==2.4.0
flake8-docstrings==0.2.1.post1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
mccabe==0.3
openapi-codec==1.3.2
pep257==0.5.0
pep8==1.5.7
psycopg2==2.6
pyflakes==0.8.1
Pygments==2.2.0
requests==2.18.4
simplejson==3.11.1
six==1.9.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.11
XlsxWriter==1.0.0
xlwt==1.0.0
